feat: 이미지 업로드 모킹 추가 및 테스트 수정 #21
